requirements.txt
asgiref==3.5.2
Django==4.0.5
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.8.4
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
